Refactoring to keep things neat
