+++ frontend/src/components/Images/index.js
@@ -1,0 +1,3 @@
+import ImageUpload from "./ImageUpload";
+
+export { ImageUpload };
